+++ Sources/StreamChat/Database/DatabaseSession.swift
@@ -393,6 +393,7 @@ protocol AttachmentDatabaseSession {
 }
 
 protocol QueuedRequestDatabaseSession {
+    func allQueuedRequests() -> [QueuedRequestDTO]
     func deleteQueuedRequest(id: String)
 }